+++ packages/camera/camera/example/lib/main.dart
@@ -127,6 +127,7 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
       _initializeCameraController(cameraController.description);
     }
   }
+
   // #enddocregion AppLifecycle
 
   @override